Implement Pruning Search Approach
Consider the search space to have an upper bound of f^b possible solutions (b decisions to send any given firefighter there).
However, when firefighters are on the same source space, it doesn't matter which firefighter you send.
Additionally, since dispatching firefighters can only add to the distance traveled, as soon as we have a complete solution,
all possibilities already as long as the known shortest solution can be discarded. Thus, the actual runtime
depends on how easily pruned the possibility space is.